+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added set methods to `ListExpr` and `ListNameSpace`
 - Added `filter` method to `ListExpr` and `ListNameSpace`
 - Added `extra_columns` option to `scan_parquet` method
+- Removed `columns` option from `DataFrame` constructor (use `schema` instead)
 
 ## 0.20.0 (2025-06-23)